Subject: shop and cart functionality
body:
getting the user to be able to ad items to cart with
a quantity which is added to the object.
the cart now takes the quantity and uses that to
display how many items are in the cart.
The nav bar is now sticky which help to show when things
are added to the basket. maybe an animation would be nice.

Need to sort styling on the shop and cart but functionality
is pretty much there.