Subject: shoppingCart editing quantity
Body:
Applying the quantity function to handle the user changing
the amount of items, displaying an error if the user
inputs less than 1.
restyling to show more central and giving it more padding
and space.

need to add a burger menu and finish styling the total
amount for the cart