Subject: Nav Mobile
Body:
Applied the correct functions and styling to mobile nav and using a
toggle to toggle the styles on and off was a lot more efficient than
creating 2 functions, using state was the best approach. maybe some
styling needs to be done to the burger button but its functional and
looks okay.