Subject: Finish index, footer and shoppingCart
Body:
Making a couple of adjustments to the index page adding a price to the
best sellers and changing a couple of styles.
Adding a generic footer that displays on all pages and has non-working
links but mainly for aesthetics.
Changing the cart total and adding a number of items piece to show how
many items are in the cart. Adding some styling to wrap it all off.